Code review
Minor factorizations.
Fixed comments.
Moved ND2 specific code to a helper file.
